refactor(admin): split session actions
Move login and logout implementation details into an app-local session helper so actions.ts stays focused on the exported server-action surface while auth behavior remains unchanged.

Verified with pnpm exec prettier --check apps/admin/src/app/actions.ts apps/admin/src/app/lib/session.ts; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; git diff --check -- apps/admin/src/app/actions.ts apps/admin/src/app/lib/session.ts; pnpm validate.
