+++ apps/admin/src/app/actions.ts
@@ -1,17 +1,5 @@
 "use server";
 
-import { cookies } from "next/headers";
-import {
-  verifyAdminPassword,
-  verifyAdminTotp,
-  createSessionToken,
-  ADMIN_COOKIE,
-  ADMIN_COOKIE_OPTIONS,
-} from "@anipotts/lib/admin";
-import { adminLoginSchema, formatZodError } from "@anipotts/lib/validation";
-import { getEnv } from "@anipotts/lib/env";
-import { checkAdminLoginRateLimit } from "@/lib/rateLimit";
-import { headers } from "next/headers";
 import type { ContentStatus, Platform, VoiceMode } from "@anipotts/types";
 import { withAuth } from "./lib/with-auth";
 import {
@@ -42,6 +30,7 @@ import {
   saveProjectDraft,
   saveWritingDraft,
 } from "./lib/site-content";
+import { loginAdmin, logoutAdmin } from "./lib/session";
 import type {
   CmsProjectContent,
   CmsWritingContent,
@@ -57,59 +46,11 @@ const SAFE_EXTERNAL_ID_RE = /^[a-zA-Z0-9_-]{1,128}$/;
 // ── Auth ──
 
 export async function login(formData: FormData) {
-  const hdrs = await headers();
-  const ip =
-    hdrs.get("x-forwarded-for")?.split(",")[0]?.trim() ||
-    hdrs.get("x-real-ip") ||
-    "unknown";
-  const rateLimit = await checkAdminLoginRateLimit(ip);
-  if (!rateLimit.success) {
-    return { error: "Too many login attempts. Try again later." };
-  }
-
-  const raw = {
-    password: formData.get("password") as string,
-    totp: (formData.get("totp") as string) || "",
-  };
-
-  const parsed = adminLoginSchema.safeParse(raw);
-  if (!parsed.success) {
-    return formatZodError(parsed.error);
-  }
-
-  const { password, totp } = parsed.data;
-
-  const pwResult = verifyAdminPassword(password, getEnv("ADMIN_PASSWORD"));
-  if (!pwResult.success) {
-    return { error: pwResult.error || "Invalid password" };
-  }
-
-  const totpSecret = getEnv("ADMIN_TOTP_SECRET");
-  if (totpSecret) {
-    if (!totp) {
-      return { error: "TOTP code is required" };
-    }
-    const totpResult = verifyAdminTotp(totp, totpSecret);
-    if (!totpResult.success) {
-      return { error: totpResult.error || "Invalid TOTP" };
-    }
-  }
-
-  const secret = getEnv("ADMIN_PASSWORD") ?? "";
-  const token = createSessionToken(secret);
-  const jar = await cookies();
-  jar.set(ADMIN_COOKIE, token, {
-    ...ADMIN_COOKIE_OPTIONS,
-    sameSite: ADMIN_COOKIE_OPTIONS.sameSite as "strict" | "lax" | "none",
-  });
-
-  return { success: true };
+  return loginAdmin(formData);
 }
 
 export async function logout() {
-  const jar = await cookies();
-  jar.delete(ADMIN_COOKIE);
-  return { success: true };
+  return logoutAdmin();
 }
 
 // ── Site Copy ──
